requirements.txt
fastapi
uvicorn[standard]
squareup
sqlalchemy
python-multipart
Pillow
alembic
python-dotenv